Move to new codebase with typescript
... and done lots of things
that I lost track of which is which
and I really don't know how to use version control,
or how to vanish the scourge of abominations that
keep piling in the project. I really should just
create a new repository but what the heck...
In the end, I didn't wasn't able to port my code
to typescript because I'm dumb and used an unconvetional
coding style and converting them to typescript seems
like ton of added work.